more util functions and move read file route
